# apps/web/src/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorizationRequest": {
        "properties": {
          "authorized_by": {
            "title": "Authorized By",
            "type": "string"
          },
          "evidence_reference": {
            "title": "Evidence Reference",
            "type": "string"
          },
          "granted_at": {
            "title": "Granted At",
            "type": "string"
          },
          "granted_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Granted To"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          }
        },
        "required": [
          "authorized_by",
          "evidence_reference",
          "permissions",
          "granted_at"
        ],
        "title": "AuthorizationRequest",
        "type": "object"
      },
...
      "FilingEventRequest": {
        "properties": {
          "event_type": {
            "pattern": "^(FILED_EXTERNALLY|PAYMENT_RECORDED|E_VERIFIED|REJECTED|ACKNOWLEDGEMENT_UPLOADED)$",
            "title": "Event Type",
            "type": "string"
          },
          "evidence_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Reference"
          },
          "occurred_at": {
            "title": "Occurred At",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          }
        },
        "required": [
          "event_type",
          "reference",
          "occurred_at"
        ],
        "title": "FilingEventRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScheduleLine": {
        "properties": {
          "amount_paise": {
            "title": "Amount Paise",
            "type": "integer"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "line_key": {
            "title": "Line Key",
            "type": "string"
          },
          "reviewer_adjustment_paise": {
            "default": 0,
            "title": "Reviewer Adjustment Paise",
            "type": "integer"
          },
          "rounding_rule": {
            "default": "NEAREST_RUPEE",
            "title": "Rounding Rule",
            "type": "string"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "schedule_key": {
            "title": "Schedule Key",
            "type": "string"
          },
          "source_links": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Source Links",
            "type": "array"
          }
        },
        "required": [
          "schedule_key",
          "line_key",
          "amount_paise",
          "source_links",
          "explanation"
        ],
        "title": "ScheduleLine",
        "type": "object"
      },
...
      "SourceRequest": {
        "properties": {
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "extracted_values": {
            "additionalProperties": true,
            "title": "Extracted Values",
            "type": "object"
          },
          "source_period": {
            "title": "Source Period",
            "type": "string"
          },
          "source_type": {
            "pattern": "^(AIS|TIS|26AS|FORM16|FORM16A|BOOKS|BANK|EVIDENCE)$",
            "title": "Source Type",
            "type": "string"
          },
          "source_version": {
            "title": "Source Version",
            "type": "string"
          },
          "taxpayer_pan": {
            "pattern": "^[A-Z]{5}[0-9]{4}[A-Z]$",
            "title": "Taxpayer Pan",
            "type": "string"
          }
        },
        "required": [
          "source_type",
          "source_period",
          "taxpayer_pan",
          "source_version"
        ],
        "title": "SourceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceRequest": {
        "properties": {
          "assessment_year": {
            "pattern": "^[0-9]{4}-[0-9]{2}$",
            "title": "Assessment Year",
            "type": "string"
          },
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "taxpayer_name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Taxpayer Name",
            "type": "string"
          },
          "taxpayer_pan": {
            "pattern": "^[A-Z]{5}[0-9]{4}[A-Z]$",
            "title": "Taxpayer Pan",
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "assessment_year",
          "taxpayer_pan",
          "taxpayer_name"
        ],
        "title": "WorkspaceRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "firmOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/itr/workspaces": {
      "post": {
        "operationId": "create_api_itr_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
    "/api/itr/workspaces/latest": {
      "get": {
        "operationId": "latest_api_itr_workspaces_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "assessment_year",
            "required": true,
            "schema": {
              "title": "Assessment Year",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
    "/api/itr/workspaces/{workspace_id}/approve": {
      "post": {
        "operationId": "approve_api_itr_workspaces__workspace_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
    "/api/itr/workspaces/{workspace_id}/authorizations": {
      "post": {
        "operationId": "authorize_api_itr_workspaces__workspace_id__authorizations_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authorize",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
...
    "/api/itr/workspaces/{workspace_id}/events": {
      "post": {
        "operationId": "event_api_itr_workspaces__workspace_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilingEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
    "/api/itr/workspaces/{workspace_id}/pack": {
      "post": {
        "operationId": "pack_api_itr_workspaces__workspace_id__pack_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pack",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
    "/api/itr/workspaces/{workspace_id}/reconcile": {
      "post": {
        "operationId": "reconcile_api_itr_workspaces__workspace_id__reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
    "/api/itr/workspaces/{workspace_id}/schedules": {
      "put": {
        "operationId": "schedules_api_itr_workspaces__workspace_id__schedules_put",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/ScheduleLine"
                },
                "title": "Lines",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedules",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
    "/api/itr/workspaces/{workspace_id}/sources": {
      "post": {
        "operationId": "add_source_api_itr_workspaces__workspace_id__sources_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Source",
        "tags": [
          "itr-workspaces"
        ]
      }
    },
...
  }
}